Wrote Irrization round logic

+++ Municipality/Entities/AgriculturalLand.cs
@@ -6,7 +6,7 @@ public class AgriculturalLand
 {
     [Key]
     public int LandQueueNumber { get; set; }
-    public double Area { get; set; }
+    public decimal Area { get; set; }
     public int AgricultureAreaId { get; set; }
     public string? OwnerPin { get; set; }
 }
